+++ packages/remote-react-components/src/auto-generated/SuspenseTrigger.ts
@@ -1,0 +1,21 @@
+/* prettier-ignore */
+/* This file is auto-generated with the remote-components-generator */
+"use client";
+import createFlowRemoteComponent from "@/components/createFlowRemoteComponent";
+import { RemoteSuspenseTriggerElement } from "@mittwald/flow-remote-elements";
+export { type RemoteSuspenseTriggerElement } from "@mittwald/flow-remote-elements";
+
+export const SuspenseTrigger = createFlowRemoteComponent(
+  "flr-suspense-trigger",
+  "SuspenseTrigger",
+  {
+    clearPropsContext: false,
+  },
+  RemoteSuspenseTriggerElement,
+  {
+    slotProps: {
+      wrapper: "flr-slot-root-wrapper",
+    },
+    eventProps: {},
+  },
+);